+++ inject/inject-modules/src/main/scala/com/twitter/inject/modules/LoggerModule.scala
@@ -16,5 +16,5 @@ object LoggerModule extends TwitterModule {
   }
 
   /**  Java-friendly way to access this module as a singleton instance */
-  def apply(): this.type = this
+  def get(): this.type = this
 }