+++ inject/inject-modules/src/main/scala/com/twitter/inject/modules/StackTransformerModule.scala
@@ -23,4 +23,7 @@ object StackTransformerModule extends TwitterModule {
       ): Stack[ServiceFactory[Req, Rep]] =
         transformers.foldLeft(stackSvcFac)((stack, transform) => transform(stack))
     }
+
+  /**  Java-friendly way to access this module as a singleton instance */
+  def get(): this.type = this
 }