+++ inject/inject-thrift-client/src/main/scala/com/twitter/inject/thrift/modules/AndThenServiceModule.scala
@@ -9,4 +9,7 @@ object AndThenServiceModule extends TwitterModule {
   override def configure: Unit = {
     bind[AndThenService].to[DefaultAndThenServiceImpl]
   }
+
+  /**  Java-friendly way to access this module as a singleton instance */
+  def get(): this.type = this
 }